[binance] Move exchangeInfo method to MarketDataService
