+++ xchange-binance/src/main/java/org/knowm/xchange/binance/service/BinanceBaseService.java
@@ -1,14 +1,11 @@
 package org.knowm.xchange.binance.service;
 
-import static org.knowm.xchange.binance.BinanceResilience.REQUEST_WEIGHT_RATE_LIMITER;
-
 import java.io.IOException;
 import org.knowm.xchange.ExchangeSpecification;
 import org.knowm.xchange.binance.BinanceAuthenticated;
 import org.knowm.xchange.binance.BinanceExchange;
 import org.knowm.xchange.binance.BinanceFuturesAuthenticated;
 import org.knowm.xchange.binance.dto.meta.BinanceSystemStatus;
-import org.knowm.xchange.binance.dto.meta.exchangeinfo.BinanceExchangeInfo;
 import org.knowm.xchange.client.ExchangeRestProxyBuilder;
 import org.knowm.xchange.client.ResilienceRegistries;
 import org.knowm.xchange.service.BaseResilientExchangeService;
@@ -90,20 +87,6 @@ public class BinanceBaseService extends BaseResilientExchangeService<BinanceExch
 
   public SynchronizedValueFactory<Long> getTimestampFactory() {
     return exchange.getTimestampFactory();
-  }
-
-  public BinanceExchangeInfo getExchangeInfo() throws IOException {
-    return decorateApiCall(binance::exchangeInfo)
-        .withRetry(retry("exchangeInfo"))
-        .withRateLimiter(rateLimiter(REQUEST_WEIGHT_RATE_LIMITER))
-        .call();
-  }
-
-  public BinanceExchangeInfo getFutureExchangeInfo() throws IOException {
-    return decorateApiCall(binanceFutures::exchangeInfo)
-        .withRetry(retry("exchangeInfo"))
-        .withRateLimiter(rateLimiter(REQUEST_WEIGHT_RATE_LIMITER))
-        .call();
   }
 
   public BinanceSystemStatus getSystemStatus() throws IOException {
